app/core/models.py
from django.conf import settings
from django.db import models
from core.util import phone_regex, validate_deadline
from django.contrib.auth.models import (
    AbstractBaseUser,
    BaseUserManager,
    PermissionsMixin
)

class UserManager(BaseUserManager):
    def create_user(self, email, password=None, **extra_fields):
        if not email:
            raise ValueError("Users must have an email address")

        user = self.model(email=self.normalize_email(email), **extra_fields)
        user.set_password(password)

        user.save(using=self._db)

        return user

    def create_superuser(self, email, password):
        user = self.create_user(email, password)
        user.is_staff = True
        user.is_superuser = True
        user.save(using=self._db)

        return user

class User(AbstractBaseUser, PermissionsMixin):
    email = models.EmailField(max_length=255, unique=True)
    name = models.CharField(max_length=255)

    is_active = models.BooleanField(default=True)
    is_staff = models.BooleanField(default=False)

    objects = UserManager()

    USERNAME_FIELD = "email"


class Category(models.Model):
    name = models.CharField(max_length=255)

    def __str__(self):
        return self.name


class Order(models.Model):
    user = models.ForeignKey(
        settings.AUTH_USER_MODEL,
        on_delete=models.CASCADE,
    )
    contact_name = models.CharField(max_length=255)
    contact_phone = models.CharField(validators=[phone_regex], max_length=17)
    description = models.TextField()
    real_state_agency = models.CharField(max_length=255)
    company = models.CharField(max_length=255)
    deadline = models.DateField(validators=[validate_deadline])
    category = models.ForeignKey(Category, on_delete=models.CASCADE)

    def __str__(self):
        return self.description